feat(dateformat-converter): module rebuilt to implement loadConfig and redo convert
to provide a custom configuration the user now needs to load the configuration first and then use
it. the convert method will now only receive the configuration names. The Convert module was dropped
and the required functionality was brought into dateformat-converter

BREAKING CHANGE: Configurations can no longer be provided to the convert method. They must first be
loaded and then the name is passed to the convert method